More double-precision updates.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@19090 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Plotter/avtTransparencyActor.h
@@ -194,7 +194,7 @@ class PLOTTER_API avtTransparencyActor
     bool                                             cachedTransparencies;
 
     void                                             SetUpActor(void);
-    void                                             PrepareDataset(int, int);
+    void                                             PrepareDataset(size_t, size_t);
     void                                             DetermineTransparencies();
 };
 #endif
